Add VueJS and make it compile
