+++ packages/server-web/package.json
@@ -23,6 +23,8 @@
     "jest": "^22.4.4",
     "prettier": "^1.12.1",
     "rollup": "^0.60.7",
+    "rollup-plugin-node-resolve": "^3.3.0",
+    "rollup-plugin-replace": "^2.0.0",
     "rollup-plugin-vue": "^4.2.0",
     "vue-template-compiler": "^2.5.16"
   },
