+++ packages/server-web/rollup.config.js
@@ -1,4 +1,8 @@
-import VuePlugin from "rollup-plugin-vue";
+import resolve from "rollup-plugin-node-resolve";
+import replace from "rollup-plugin-replace";
+import vue from "rollup-plugin-vue";
+
+const isProduction = !process.env.ROLLUP_WATCH;
 
 export default {
   input: "src/client/client.js",
@@ -6,5 +10,17 @@ export default {
     format: "esm",
     file: "src/dist/client.dist.js"
   },
-  plugins: [VuePlugin(/* VuePluginOptions */)]
+  plugins: [
+    resolve(),
+    replace({
+      "process.env.NODE_ENV": JSON.stringify("production")
+    }),
+    vue({
+      template: {
+        isProduction,
+        compilerOptions: { preserveWhitespace: false }
+      },
+      css: true
+    })
+  ]
 };
