+++ packages/server-web/src/client/client.js
@@ -1,12 +1,5 @@
-import io from "./socket.io-bundle.js";
+import createSocketConnection from "./socket-connection.js";
+import createApp from "./create-app.js";
 
-const socket = io();
-
-socket.on("connection", events => events.forEach(processEvent));
-socket.on("event", processEvent);
-
-socket.emit("command", { command: "ADD_TODO", data: { title: "create a nice test todo" } });
-
-function processEvent(event) {
-  console.log(new Date(), "processing event", event);
-}
+createSocketConnection();
+createApp();
