+++ packages/server-web/src/client/components/Home.vue
@@ -1,0 +1,17 @@
+<template>
+  <div>
+    <h1>Todastic</h1>
+    <p>the musical - of Vue!</p>
+  </div>
+</template>
+
+<script>
+console.log("todastic of vue rendering");
+export default {};
+</script>
+
+<style>
+div {
+  background-color: red;
+}
+</style>
